+++ app/src/main/java/pl/lambada/songsync/util/LyricsUtils.kt
@@ -1,0 +1,133 @@
+package pl.lambada.songsync.util
+
+import android.annotation.SuppressLint
+import android.app.PendingIntent
+import android.app.RecoverableSecurityException
+import android.content.Context
+import android.net.Uri
+import android.os.Build
+import android.os.Environment
+import android.os.ParcelFileDescriptor
+import android.provider.MediaStore
+import android.util.Log
+import androidx.documentfile.provider.DocumentFile
+import com.kyant.taglib.TagLib
+import pl.lambada.songsync.domain.model.SongInfo
+import pl.lambada.songsync.util.ext.toLrcFile
+import java.io.File
+import java.io.FileNotFoundException
+
+fun generateLrcContent(
+    song: SongInfo,
+    lyrics: String,
+    generatedUsingString: String
+): String {
+    return "[ti:${song.songName}]\n[ar:${song.artistName}]\n[by:$generatedUsingString]\n$lyrics"
+}
+
+fun newLyricsFilePath(filePath: String?, song: SongInfo): File {
+    return filePath?.toLrcFile() ?: File(
+        Environment.getExternalStoragePublicDirectory(Environment.DIRECTORY_DOWNLOADS),
+        "SongSync/${song.songName} - ${song.artistName}.lrc"
+    )
+}
+
+@SuppressLint("Range")
+fun getFileDescriptorFromPath(
+    context: Context, filePath: String, mode: String = "r"
+): ParcelFileDescriptor? {
+    val resolver = context.contentResolver
+    val uri = MediaStore.Audio.Media.EXTERNAL_CONTENT_URI
+
+    val projection = arrayOf(MediaStore.Files.FileColumns._ID)
+    val selection = "${MediaStore.Files.FileColumns.DATA}=?"
+    val selectionArgs = arrayOf(filePath)
+
+    return resolver.query(uri, projection, selection, selectionArgs, null)?.use { cursor ->
+        if (cursor.moveToFirst()) {
+            val fileId = cursor.getInt(cursor.getColumnIndex(MediaStore.Files.FileColumns._ID))
+            if (fileId != -1) {
+                val fileUri = Uri.withAppendedPath(uri, fileId.toString())
+                try {
+                    resolver.openFileDescriptor(fileUri, mode)
+                } catch (e: FileNotFoundException) {
+                    Log.e("LyricsFetchViewModel", "File not found: ${e.message}")
+                    null
+                }
+            } else null
+        } else null
+    }
+}
+
+fun embedLyricsInFile(
+    context: Context,
+    filePath: String,
+    lyrics: String,
+    securityExceptionHandler: (PendingIntent) -> Unit = {}
+): Boolean {
+    return try {
+        val fd = getFileDescriptorFromPath(context, filePath, mode = "w")
+            ?: throw IllegalStateException("File descriptor is null")
+
+        val fileDescriptor = fd.dup().detachFd()
+        val metadata = TagLib.getMetadata(fileDescriptor, false) ?: error("Metadata is null")
+
+        TagLib.savePropertyMap(
+            fd.dup().detachFd(),
+            propertyMap = metadata.propertyMap.apply { put("LYRICS", arrayOf(lyrics)) }
+        )
+
+        true
+    } catch (securityException: SecurityException) {
+        handleSecurityException(securityException, securityExceptionHandler)
+        false
+    } catch (e: Exception) {
+        Log.e("LyricsFetchViewModel", "Error embedding lyrics: ${e.message}")
+        false
+    }
+}
+
+fun handleSecurityException(
+    securityException: SecurityException,
+    intentPassthrough: (PendingIntent) -> Unit
+) {
+    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.Q) {
+        val recoverableSecurityException =
+            securityException as? RecoverableSecurityException
+                ?: throw RuntimeException(securityException.message, securityException)
+
+        intentPassthrough(recoverableSecurityException.userAction.actionIntent)
+    } else {
+        throw RuntimeException(securityException.message, securityException)
+    }
+}
+
+
+
+
+fun saveToExternalPath(
+    context: Context,
+    sourceFilePath: String?,
+    lrc: String,
+    fileName: String,
+    newLyricsFilePath: String?
+) {
+    val sd = context.externalCacheDirs[1].absolutePath.substringBefore("/Android/data")
+    val path = sourceFilePath
+        ?.toLrcFile()
+        ?.absolutePath
+        ?.substringAfter(sd)
+        ?.split("/")
+        ?.dropLast(1)
+        ?: error("path was null when trying to save to sd card")
+    var sdCardFiles = DocumentFile.fromTreeUri(context, Uri.parse(newLyricsFilePath))
+    path.forEach { element ->
+        sdCardFiles = sdCardFiles?.listFiles()?.firstOrNull { it.name == element }
+    }
+    sdCardFiles?.listFiles()?.firstOrNull { it.name == fileName }?.delete()
+    sdCardFiles?.createFile("text/lrc", fileName)?.let {
+        context.contentResolver.openOutputStream(it.uri)?.use { outputStream ->
+            outputStream.write(lrc.toByteArray())
+        }
+    }
+}
